segment lib: support for more than 2 billion cells
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@51482 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/segment/get.c
@@ -38,7 +38,7 @@
  * \return -1 if unable to seek or read segment file
  */
 
-int segment_get(SEGMENT * SEG, void *buf, int row, int col)
+int segment_get(SEGMENT * SEG, void *buf, off_t row, off_t col)
 {
     int index, n, i;
